@@ -25,8 +25,8 @@ class AddToCart extends StatelessWidget {
         },
         style: ButtonStyle(
             backgroundColor:
-            MaterialStateProperty.all(context.theme.buttonColor),
+            MaterialStateProperty.all(context.theme.colorScheme.onPrimary),
             shape: MaterialStateProperty.all(const StadiumBorder())),
-        child: isInCart ? const Icon(Icons.done) : const Icon(CupertinoIcons.cart_badge_plus));
+        child: isInCart ? const Icon(Icons.done,color: Colors.white,) : const Icon(CupertinoIcons.cart_badge_plus,color: Colors.white,));
   }
 }
